requirements.txt
requests==2.28.1
mysql-connector-python==8.0.30
python-dotenv==0.21.0
